refactor: move getTravelTime out of TrainrunSectionViewObject
