+++ src/app/view/editor-main-view/data-views/trainrunSectionViewObject.ts
@@ -10,6 +10,7 @@ import {EditorView} from "./editor.view";
 import {TrainrunSectionsView} from "./trainrunsections.view";
 import {Node} from "src/app/models/node.model";
 import {SHOW_MAX_SINGLE_TRAINRUN_SECTIONS_STOPS} from "../../rastering/definitions";
+import {TrainrunsectionHelper} from "src/app/services/util/trainrunsection.helper";
 
 export class TrainrunSectionViewObject {
   readonly firstSection: TrainrunSection;
@@ -74,49 +75,12 @@ export class TrainrunSectionViewObject {
     }
     return this.getCollapsedStopNodeFromStopIndex(stopIndex);
   }
-
   getTravelTime(): number {
-    if (this.trainrunSections.length === 1) {
-      return this.firstSection.getTravelTime();
-    }
-
-    return this.trainrunSections.reduce((sum, section, index) => {
-      let sectionTime = section.getTravelTime();
-
-      // Add stop time at intermediate nodes (all except the last section)
-      if (index < this.trainrunSections.length - 1) {
-        const nextSection = this.trainrunSections[index + 1];
-        const stopTime = Math.abs(
-          nextSection.getSourceDepartureConsecutiveTime() -
-            section.getTargetArrivalConsecutiveTime(),
-        );
-        sectionTime += stopTime;
-      }
-
-      return sum + sectionTime;
-    }, 0);
+    return TrainrunsectionHelper.getTravelTimeForSectionGroup(this.trainrunSections);
   }
 
   getBackwardTravelTime(): number {
-    if (this.trainrunSections.length === 1) {
-      return this.firstSection.getBackwardTravelTime();
-    }
-
-    return this.trainrunSections.reduce((sum, section, index) => {
-      let sectionTime = section.getBackwardTravelTime();
-
-      // Add stop time at intermediate nodes (all except the last section)
-      if (index < this.trainrunSections.length - 1) {
-        const nextSection = this.trainrunSections[index + 1];
-        const stopTime = Math.abs(
-          section.getTargetDepartureConsecutiveTime() -
-            nextSection.getSourceArrivalConsecutiveTime(),
-        );
-        sectionTime += stopTime;
-      }
-
-      return sum + sectionTime;
-    }, 0);
+    return TrainrunsectionHelper.getBackwardTravelTimeForSectionGroup(this.trainrunSections);
   }
 
   areTravelTimesEqual(): boolean {
